# functions/dist/swagger.yaml
components:
    examples: {}
    headers: {}
    parameters: {}
    requestBodies: {}
    responses: {}
    schemas:
        StatusModel:
            enum:
                - Stripe_waiting
                - Active
                - Inactive
                - Stripe_inactive
            type: string
        MerchantOutput:
            properties:
                drveUid:
                    type: string
                shopifyDomain:
                    type: string
                commissionPercentage:
                    type: number
                    format: double
                enabled:
                    type: boolean
                name:
                    type: string
                chargeWebhookUrl:
                    type: string
                connectedAccountWebhookUrl:
                    type: string
                stripeOnConnectionRedirectURL:
                    type: string
                id:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                status:
                    $ref: '#/components/schemas/StatusModel'
                gatewayURL:
                    type: string
                gatewayAccountId:
                    type: string
                gatewayCredentials:
                    type: string
                stripeAcctId:
                    type: string
                stripeURLToConnectAccount:
                    type: string
                testMode:
                    type: boolean
            required:
                - drveUid
                - shopifyDomain
                - commissionPercentage
                - enabled
                - name
                - chargeWebhookUrl
                - connectedAccountWebhookUrl
                - stripeOnConnectionRedirectURL
                - id
                - createdAt
                - updatedAt
                - status
                - gatewayURL
                - gatewayAccountId
                - gatewayCredentials
                - stripeAcctId
                - stripeURLToConnectAccount
                - testMode
            type: object
            additionalProperties: true
        Partial_MerchantOutput_:
            properties:
                status:
                    $ref: '#/components/schemas/StatusModel'
                gatewayURL:
                    type: string
                gatewayAccountId:
                    type: string
                gatewayCredentials:
                    type: string
                stripeAcctId:
                    type: string
                stripeURLToConnectAccount:
                    type: string
                testMode:
                    type: boolean
                drveUid:
                    type: string
                shopifyDomain:
                    type: string
                commissionPercentage:
                    type: number
                    format: double
                enabled:
                    type: boolean
                name:
                    type: string
                chargeWebhookUrl:
                    type: string
                connectedAccountWebhookUrl:
                    type: string
                stripeOnConnectionRedirectURL:
                    type: string
                id:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
            type: object
            description: 'Make all properties in T optional'
        MerchantInput:
            properties:
                drveUid:
                    type: string
                shopifyDomain:
                    type: string
                commissionPercentage:
                    type: number
                    format: double
                enabled:
                    type: boolean
                name:
                    type: string
                chargeWebhookUrl:
                    type: string
                connectedAccountWebhookUrl:
                    type: string
                stripeOnConnectionRedirectURL:
                    type: string
            required:
                - drveUid
                - shopifyDomain
                - commissionPercentage
                - enabled
                - name
                - chargeWebhookUrl
                - connectedAccountWebhookUrl
                - stripeOnConnectionRedirectURL
            type: object
            additionalProperties: true
    securitySchemes:
        api_key:
            scheme: basic
            type: http
info:
    title: functions
    version: 1-alpha
openapi: 3.0.0
paths:
    /merchants:
        get:
            operationId: GetMerchants
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/MerchantOutput'
                                type: array
                    description: Ok
            security:
                -
                    api_key: []
            parameters: []
        post:
            operationId: CreateMerchant
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MerchantOutput'
                    description: Ok
            security:
                -
                    api_key: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MerchantInput'
    '/merchants/{drveUid}':
        get:
            operationId: GetMerchant
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Partial_MerchantOutput_'
                    description: Ok
            security:
                -
                    api_key: []
            parameters:
                -
                    in: path
                    name: drveUid
                    required: true
                    schema:
                        type: string
        put:
            operationId: EditMerchant
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MerchantOutput'
                    description: Ok
            security:
                -
                    api_key: []
            parameters:
                -
                    in: path
                    name: drveUid
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/MerchantInput'
    /merchantsCreate:
        get:
            operationId: CreateMerchants
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                additionalProperties: true
                                type: object
                    description: Ok
            security: []
            parameters: []
    /merchantsCreate/nestedTest:
        get:
            operationId: CreateNestedExample
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                additionalProperties: true
                                type: object
                    description: Ok
            security: []
            parameters: []
servers:
    -
        url: 'https://localhost:3000/v1'